+++ module/rest/src/symbols.ts
